refactored ui code to windows module
